Moving staticfiles_compat to utils.
